+++ core/src/main/java/com/fpf/smartscansdk/core/data/Processors.kt
@@ -1,15 +1,5 @@
 package com.fpf.smartscansdk.core.data
 
-import android.content.Context
-
-interface IProcessorListener<Input, Output> {
-    suspend fun onActive(context: Context) = Unit
-    suspend fun onBatchComplete(context: Context, batch: List<Output>) = Unit
-    suspend fun onComplete(context: Context, metrics: Metrics.Success) = Unit
-    suspend fun onProgress(context: Context, progress: Float) = Unit
-    fun onError(context: Context, error: Exception, item: Input) = Unit
-    suspend fun onFail(context: Context, failureMetrics: Metrics.Failure) = Unit
-}
 
 sealed class Metrics {
     data class Success(val totalProcessed: Int = 0, val timeElapsed: Long = 0L) : Metrics()